configure network with net_helper
- removes the sp_overwrite_iface_conf option and the direct use
of iface_genconf
- removes sp_single_iface, for all such cases use a secondary
interface, even if it is down.
- removes bond adjustments, turns out the only relevant piece there
is already happening in configure_network_redhat.yml
(ifup-pre-local for RHEL based)
- adds all relevant configuration parameters for net_helper
genconfig
- adds a task that applies configuration with net_helper
applyifcfg
- checks whether a previous configuration exists, complains if
so (the only way to remove a config is to manually execute
net_helper cleanup on the node, i.e. this is not yet added to
the playbook)